+++ assets/version_check.py
@@ -5,7 +5,7 @@ import sys
 import codecs
 
 Failed = 0
-VERSION = "0.2"
+VERSION = "0.3"
 
 INDEX_ITEMS = ['<p>Version {0}</p>']
 CHANGELOG_ITEMS = [