+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3] - 2026-02-09
 ### Added
 - Recently visited section
 - `renderRecent` function
@@ -24,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Redirect page
 
 
-[Unreleased]: https://github.com/sepandhaghighi/githubp/compare/v0.2...main
+[Unreleased]: https://github.com/sepandhaghighi/githubp/compare/v0.3...main
+[0.3]: https://github.com/sepandhaghighi/githubp/compare/v0.2...v0.3
 [0.2]: https://github.com/sepandhaghighi/githubp/compare/v0.1...v0.2
 [0.1]: https://github.com/sepandhaghighi/githubp/compare/c80d750...v0.1